backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9

alembic==1.13.1
python-dotenv==1.0.1

pydantic==2.7.4
openai==1.109.1
strawberry-graphql[fastapi]==0.224.0

pyjwt==2.8.0
passlib[bcrypt]==1.7.4

pinecone-client==4.1.0
pypdf==5.1.0
python-multipart==0.0.12
aiofiles==23.2.1
